+++ core/src/main/java/com/graphhopper/routing/ev/BooleanEncodedValue.java
@@ -1,21 +1,10 @@
 package com.graphhopper.routing.ev;
 
-import com.graphhopper.storage.IntsRef;
-
 /**
  * This interface defines access to an edge property of type boolean. The default value is false.
  */
 public interface BooleanEncodedValue extends EncodedValue {
+    void setBool(boolean reverse, int edgeId, EdgeIntAccess edgeIntAccess, boolean value);
 
-    default void setBool(boolean reverse, IntsRef ref, boolean value) {
-        setBool(-1, reverse, ref, value);
-    }
-
-    void setBool(int edgeId, boolean reverse, IntsRef ref, boolean value);
-
-    default boolean getBool(boolean reverse, IntsRef ref) {
-        return getBool(-1, reverse, ref);
-    }
-
-    boolean getBool(int edgeId, boolean reverse, IntsRef ref);
+    boolean getBool(boolean reverse, int edgeId, EdgeIntAccess edgeIntAccess);
 }