+++ core/src/main/java/com/graphhopper/routing/util/parsers/OSMFootwayParser.java
@@ -21,6 +21,7 @@ package com.graphhopper.routing.util.parsers;
 import com.graphhopper.reader.ReaderWay;
 import com.graphhopper.routing.ev.EnumEncodedValue;
 import com.graphhopper.routing.ev.Footway;
+import com.graphhopper.routing.ev.EdgeIntAccess;
 import com.graphhopper.storage.IntsRef;
 
 /**
@@ -34,8 +35,8 @@ public class OSMFootwayParser implements TagParser {
     }
 
     @Override
-    public void handleWayTags(IntsRef edgeFlags, ReaderWay way, IntsRef relationFlags) {
+    public void handleWayTags(int edgeId, EdgeIntAccess edgeIntAccess, ReaderWay way, IntsRef relationFlags) {
         String footway = way.getTag("footway");
-        footwayEnc.setEnum(false, edgeFlags, Footway.find(footway));
+        footwayEnc.setEnum(false, edgeId, edgeIntAccess, Footway.find(footway));
     }
 }