+++ core/src/main/java/com/graphhopper/routing/util/parsers/OSMWayIDParser.java
@@ -19,6 +19,7 @@
 package com.graphhopper.routing.util.parsers;
 
 import com.graphhopper.reader.ReaderWay;
+import com.graphhopper.routing.ev.EdgeIntAccess;
 import com.graphhopper.routing.ev.IntEncodedValue;
 import com.graphhopper.storage.IntsRef;
 
@@ -30,12 +31,12 @@ public class OSMWayIDParser implements TagParser {
     }
 
     @Override
-    public void handleWayTags(IntsRef edgeFlags, ReaderWay way, IntsRef relationFlags) {
+    public void handleWayTags(int edgeId, EdgeIntAccess edgeIntAccess, ReaderWay way, IntsRef relationFlags) {
         if (way.getId() > osmWayIdEnc.getMaxStorableInt())
             throw new IllegalArgumentException("Cannot store OSM way ID: " + way.getId() + " as it is too large (> "
                     + osmWayIdEnc.getMaxStorableInt() + "). You can disable " + osmWayIdEnc.getName() + " if you do not " +
                     "need to store the OSM way IDs");
         int wayId = Math.toIntExact(way.getId());
-        osmWayIdEnc.setInt(false, edgeFlags, wayId);
+        osmWayIdEnc.setInt(false, edgeId, edgeIntAccess, wayId);
     }
 }