+++ core/src/test/java/com/graphhopper/routing/util/parsers/OSMTrackTypeParserTest.java
@@ -1,9 +1,7 @@
 package com.graphhopper.routing.util.parsers;
 
 import com.graphhopper.reader.ReaderWay;
-import com.graphhopper.routing.ev.EncodedValue;
-import com.graphhopper.routing.ev.EnumEncodedValue;
-import com.graphhopper.routing.ev.TrackType;
+import com.graphhopper.routing.ev.*;
 import com.graphhopper.storage.IntsRef;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -17,7 +15,7 @@ public class OSMTrackTypeParserTest {
 
     @BeforeEach
     public void setUp() {
-        ttEnc = new EnumEncodedValue<>(TrackType.KEY, TrackType.class);
+        ttEnc = TrackType.create();
         ttEnc.init(new EncodedValue.InitializerConfig());
         parser = new OSMTrackTypeParser(ttEnc);
         relFlags = new IntsRef(2);
@@ -26,46 +24,49 @@ public class OSMTrackTypeParserTest {
     @Test
     public void testSimpleTags() {
         ReaderWay readerWay = new ReaderWay(1);
-        IntsRef intsRef = new IntsRef(1);
+        EdgeIntAccess edgeIntAccess = new ArrayEdgeIntAccess(1);
+        int edgeId = 0;
         readerWay.setTag("tracktype", "grade1");
-        parser.handleWayTags(intsRef, readerWay, relFlags);
-        assertEquals(TrackType.GRADE1, ttEnc.getEnum(false, intsRef));
+        parser.handleWayTags(edgeId, edgeIntAccess, readerWay, relFlags);
+        assertEquals(TrackType.GRADE1, ttEnc.getEnum(false, edgeId, edgeIntAccess));
 
-        intsRef = new IntsRef(1);
+        edgeIntAccess = new ArrayEdgeIntAccess(1);
         readerWay.setTag("tracktype", "grade2");
-        parser.handleWayTags(intsRef, readerWay, relFlags);
-        assertEquals(TrackType.GRADE2, ttEnc.getEnum(false, intsRef));
+        parser.handleWayTags(edgeId, edgeIntAccess, readerWay, relFlags);
+        assertEquals(TrackType.GRADE2, ttEnc.getEnum(false, edgeId, edgeIntAccess));
 
-        intsRef = new IntsRef(1);
+        edgeIntAccess = new ArrayEdgeIntAccess(1);
         readerWay.setTag("tracktype", "grade3");
-        parser.handleWayTags(intsRef, readerWay, relFlags);
-        assertEquals(TrackType.GRADE3, ttEnc.getEnum(false, intsRef));
+        parser.handleWayTags(edgeId, edgeIntAccess, readerWay, relFlags);
+        assertEquals(TrackType.GRADE3, ttEnc.getEnum(false, edgeId, edgeIntAccess));
 
-        intsRef = new IntsRef(1);
+        edgeIntAccess = new ArrayEdgeIntAccess(1);
         readerWay.setTag("tracktype", "grade4");
-        parser.handleWayTags(intsRef, readerWay, relFlags);
-        assertEquals(TrackType.GRADE4, ttEnc.getEnum(false, intsRef));
+        parser.handleWayTags(edgeId, edgeIntAccess, readerWay, relFlags);
+        assertEquals(TrackType.GRADE4, ttEnc.getEnum(false, edgeId, edgeIntAccess));
 
-        intsRef = new IntsRef(1);
+        edgeIntAccess = new ArrayEdgeIntAccess(1);
         readerWay.setTag("tracktype", "grade5");
-        parser.handleWayTags(intsRef, readerWay, relFlags);
-        assertEquals(TrackType.GRADE5, ttEnc.getEnum(false, intsRef));
+        parser.handleWayTags(edgeId, edgeIntAccess, readerWay, relFlags);
+        assertEquals(TrackType.GRADE5, ttEnc.getEnum(false, edgeId, edgeIntAccess));
     }
 
     @Test
     public void testUnkownValue() {
         ReaderWay readerWay = new ReaderWay(1);
-        IntsRef intsRef = new IntsRef(1);
+        EdgeIntAccess edgeIntAccess = new ArrayEdgeIntAccess(1);
+        int edgeId = 0;
         readerWay.setTag("tracktype", "unknownstuff");
-        parser.handleWayTags(intsRef, readerWay, relFlags);
-        assertEquals(TrackType.MISSING, ttEnc.getEnum(false, intsRef));
+        parser.handleWayTags(edgeId, edgeIntAccess, readerWay, relFlags);
+        assertEquals(TrackType.MISSING, ttEnc.getEnum(false, edgeId, edgeIntAccess));
     }
 
     @Test
     public void testNoNPE() {
         ReaderWay readerWay = new ReaderWay(1);
-        IntsRef intsRef = new IntsRef(1);
-        parser.handleWayTags(intsRef, readerWay, relFlags);
-        assertEquals(TrackType.MISSING, ttEnc.getEnum(false, intsRef));
+        EdgeIntAccess edgeIntAccess = new ArrayEdgeIntAccess(1);
+        int edgeId = 0;
+        parser.handleWayTags(edgeId, edgeIntAccess, readerWay, relFlags);
+        assertEquals(TrackType.MISSING, ttEnc.getEnum(false, edgeId, edgeIntAccess));
     }
 }