+++ core/src/test/java/com/graphhopper/storage/BaseGraphWithTurnCostsTest.java
@@ -20,14 +20,14 @@ package com.graphhopper.storage;
 import com.graphhopper.routing.ev.DecimalEncodedValue;
 import com.graphhopper.routing.ev.TurnCost;
 import com.graphhopper.routing.util.EncodingManager;
-import com.graphhopper.search.EdgeKVStorage.KeyValue;
+import com.graphhopper.search.KVStorage.KeyValue;
 import com.graphhopper.util.EdgeIteratorState;
 import com.graphhopper.util.Helper;
 import org.junit.jupiter.api.Test;
 
 import java.util.Random;
 
-import static com.graphhopper.search.EdgeKVStorage.KeyValue.STREET_NAME;
+import static com.graphhopper.search.KVStorage.KeyValue.STREET_NAME;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;